CNF-23104: reorganize hwmgr-plugins/ to internal/metal3-hwmgr/
Move all metal3 hardware manager code from the top-level
hwmgr-plugins/ directory (a remnant of the multi-plugin REST API
architecture) to internal/metal3-hwmgr/.

- hwmgr-plugins/controller/utils/ → internal/metal3-hwmgr/utils/
- hwmgr-plugins/metal3/controller/ → internal/metal3-hwmgr/controller/
- hwmgr-plugins/metal3/cmd/ → internal/metal3-hwmgr/cmd/
- Deleted empty hwmgr-plugins/ tree
- Updated all import paths (20 files)
- Updated coverage thresholds and Tekton pipeline paths
