+++ internal/metal3-hwmgr/controller/inventory_test.go
