+++ internal/metal3-hwmgr/controller/inventory_types.go
