+++ internal/metal3-hwmgr/utils/allocated_node.go
