+++ internal/metal3-hwmgr/utils/conditions.go
