+++ internal/metal3-hwmgr/utils/constants.go
