+++ internal/metal3-hwmgr/utils/utils.go
