+++ internal/service/resources/collector/collector_hwplugin.go
@@ -16,7 +16,7 @@ import (
 
 	rtclient "sigs.k8s.io/controller-runtime/pkg/client"
 
-	metal3controller "github.com/openshift-kni/oran-o2ims/hwmgr-plugins/metal3/controller"
+	metal3controller "github.com/openshift-kni/oran-o2ims/internal/metal3-hwmgr/controller"
 	ctlrutils "github.com/openshift-kni/oran-o2ims/internal/controllers/utils"
 	"github.com/openshift-kni/oran-o2ims/internal/service/common/async"
 	"github.com/openshift-kni/oran-o2ims/internal/service/resources/db/models"
